Now a backup of DIP before appraisal for upload. /sharedDirectory/DIPbackups - contains the backup; when processing has finished. It also contains a list of files sent to the upload script. /sharedDirectory/watchedDirectories/uploadedDIPs/ - contains the DIP sent to the upload DIP script.
While I was working on this I had some problems with the MCP:
The article.pdf file would request write lock twice from the MCP. The second time of course caused problems.
I did some fiddling with the locks, on both the client and server, and it stopped.
Not 100% sure on this resolution.



Autoconverted from SVN (revision:899)