The problem encountered was that the transcoder and MCPClient have different flow structures, and apparently the twisted framework has not yet evolved to accomodate the transcoder top-down sturcture; therefore two sub-process launching methods are needed. The Client will use the twisted .spawnProcess method, and the transcoder will use the traditional python subprocess.
The reason the transcoder can't use the spawnProcess method is because it needs the reactor to be running. The reactor needs to be run on the main thread to handle interupts, so it can't be on a sub thread. The reactor runs a single thread, that needs to return before the spawned process will execute. This breaks the transcoder top-down programming. Note: if you try to start the reactor on a sub process without installing the signal handlers, it appears to never start - no errors.

From what I've read, I thinkt he problems were with mixing subProcess and twisted (I think it's a conflict in signal handling) and we should be able to use subProcess safely in transcoder, as it will have it's own process/memory pool, even if it is a sub process of the MCP client.

Related To Issue 525:	Normalization log for AV files doesn't show normalization tool output


Autoconverted from SVN (revision:1295)