Major debugging done on the transferD. Watches were continueing to watch after sip was moved out of sip creation area, even after they were removed with the recurisive flag set to true. Found that pyinotify was appening "-unknown-path" to event.pathname, and it was breaking the recursive functionality. Wrote a custom function for dealing with this, and recursively removing the watched directories.
In order to find out what was going on, I had to recreate my archivematica user as a normal user. Log in as archivematica. Setup eclipse to work with python and the archivematica user, run the MCP in debug mode.
Warning: If you do this, the debugging breaks can mess with the timers in the transferD


Autoconverted from SVN (revision:1672)